fix: Fixed problem wit create and update (film and user)

+++ src/main/java/ru/yandex/practicum/filmorate/controller/IdGenerator.java
@@ -1,5 +1,5 @@
 package ru.yandex.practicum.filmorate.controller;
 
 interface IdGenerator {
-    public long getNewId();
+    public Integer getNewId();
 }
